4.0.0-rc12.15
* Not Fixed: Directory stuff.
* Fixed: Chat handling.

This is an initial release to add support for Twitch's changes to how chat works internally. Everything should be working. I'm not entirely happy with the code though, so I'll be reviewing this in the future when I have time to clean things up.

I haven't had a chance to go over directory features yet, and likely won't for several days because of my job.