WIP Set-Hiding
Suggestions are created every time more than 2 letters are typed after the colon, after which it updates on every new or removed letter.
Optimally we'd want some caching on the very first time we "hydrate" the emotes and attach their owner information back on afterwards.
As for filtering... That still needs to be done on the fly for obvious reasons.