4.77.12
Hey everyone, this is just a quick fix. As a result of this change, FFZ will load on embedded players, chats, and the streamer dashboard. What it won't do is load with your settings. This is due to a couple changes Twitch has made to their website that unintentionally broke the method we were using to access domain storage from alternative subdomains. Ultimately, this just means we need to prioritize getting an extension-backed storage provider setup for browser extension users and the relevant GreaseMonkey-compatible APIs for user-script users.

* Fixed: FFZ failing to load on alternative subdomains due to the Cross Origins Settings Provider breaking.